colocating types + updating paths
