+++ src/Util.ts
@@ -5,13 +5,6 @@ import {
 } from 'google-protobuf/google/protobuf/struct_pb'
 import { Timestamp } from 'google-protobuf/google/protobuf/timestamp_pb'
 import { parse as uuidParse, v4 as uuidv4 } from 'uuid'
-import type {
-  ItemData,
-  EdgeData,
-  ItemRequestErrorData,
-  GatewayRequestStatusData,
-  ReferenceData,
-} from './Types'
 import {
   CancelItemRequest,
   Edge,
@@ -28,6 +21,46 @@ import {
   ResponderState,
   Response,
 } from './__generated__/'
+
+export type EdgeData = {
+  from: ReferenceData
+  to: ReferenceData
+}
+
+export type ReferenceData = {
+  type: string
+  uniqueAttributeValue: string
+  scope: string
+}
+
+export type ItemData = {
+  type: string
+  uniqueAttribute: string
+  scope: string
+  attributes: ItemAttributes
+  metadata: Metadata | undefined
+  linkedItemRequests: ItemRequest[]
+  linkedItems: Reference[]
+}
+
+export type GatewayRequestStatusData = {
+  responderStates: Map<string, ResponderState>
+  summary: {
+    working: number
+    stalled: number
+    complete: number
+    error: number
+    cancelled: number
+    responders: number
+  }
+  postProcessingComplete: boolean
+}
+
+export type ItemRequestErrorData = {
+  scope: string
+  errorString: string
+  errorType: ItemRequestError.ErrorType
+}
 
 //
 // Private helper functions
