+++ src/index.ts
@@ -3,4 +3,3 @@ export * from './GatewaySession'
 export * from './RequestProgress'
 export * from './Responder'
 export * from './Util'
-export * from './Types'
